@@ -3,14 +3,14 @@ import { ShoppingCartService } from 'app/restaurant-detail/shopping-cart/shoppin
 import { CartItem } from 'app/restaurant-detail/shopping-cart/cart-item.model';
 import { Order } from './order.model';
 import { Observable } from 'rxjs/Observable';
-import { Http, RequestOptions, Headers } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 
 import { MEAT_API } from '../app.api';
 
 @Injectable()
 export class OrderService {
 
-    constructor(private cartService: ShoppingCartService, private http: Http) {}
+    constructor(private cartService: ShoppingCartService, private http: HttpClient) {}
 
     itemsValue(): number {
         return this.cartService.total();
@@ -37,12 +37,7 @@ export class OrderService {
     }
 
     checkOrder(order: Order): Observable<string> {
-        const headers = new Headers();
-        headers.append('Content-Type', 'application/json');
-        return this.http.post(`${MEAT_API}/orders`, 
-                              JSON.stringify(order), 
-                              new RequestOptions({headers: headers}))
-                        .map(response => response.json())
+        return this.http.post<Order>(`${MEAT_API}/orders`, order)
                         .map(order => order.id);
     }
 
